Implement update product category (#28)
* #24 editProduct.html page with failing test for ProductOptionsController

* #24  ProductOptionsControllerTest passes

* #24 Refactoring

* #24 Refactoring: Move should return products as options to ProductDaoTest

* #24 ProductCategoryOptionsController should return categories as options (passing)

* #24 UpdateProductController should update product categoryId (failing)

* #24 UpdateProductController should update product categoryId (failing)#24 UpdateProductController should update product categoryId (still failing)

* #24 UpdateProductController should update product categoryId (test passing)

* #24 UpdateProductController should return http response

* #24 UpdateProductController should redirect client to front page (test failing)

* #24 UpdateProductController should redirect client to front page (test passing)

* Resolves #24. UpdateProductController should redirect client to front page (test passing)

+++ src/main/java/no/kristiania/database/AbstractDao.java
@@ -13,7 +13,7 @@ public abstract class AbstractDao<T> {
         this.dataSource = dataSource;
     }
 
-    protected T retrieve(Long id, String sql) throws SQLException {
+    protected T retrieve(Integer id, String sql) throws SQLException {
         try (Connection connection = dataSource.getConnection()) {
             try (PreparedStatement statement = connection.prepareStatement(sql)) {
                 statement.setLong(1, id);
